Purchase vinculado con stats y review vinculado con notifications

+++ ms-stats/src/main/java/com/app/msstats/dto/UserStatsResponse.java
@@ -2,12 +2,14 @@ package com.app.msstats.dto;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record UserStatsResponse(
         Long userId,
         String nickname,
         String avatar,
         long juegosComprados,
-        double dineroGastado,
-        double promedioRatingDado
+        BigDecimal dineroGastado,
+        BigDecimal promedioRatingDado
 ) {}
